apps,lib,test: add experimental Windows support

+++ apps/btoep-add.c
@@ -65,7 +65,7 @@ int main(int argc, char** argv) {
   cmd_opts opts = {
     .on_conflict = BTOEP_CONFLICT_ERROR
   };
-  if (!opt_parse(options, 7, &opts, argc - 1, argv + 1)) {
+  if (!opt_parse(options, 7, &opts, (size_t) argc - 1, argv + 1)) {
     fprintf(stderr, "error\n");
     return B_EXIT_CODE_USAGE_ERROR;
   }
