+++ apps/btoep-find-offset.c
@@ -89,7 +89,7 @@ int main(int argc, char** argv) {
     .stop_at_data = false,
     .stop_at_missing = false
   };
-  if (!opt_parse(options, 5, &opts, argc - 1, argv + 1)) {
+  if (!opt_parse(options, 5, &opts, (size_t) argc - 1, argv + 1)) {
     fprintf(stderr, "error\n");
     return B_EXIT_CODE_USAGE_ERROR;
   }
