+++ apps/btoep-set-size.c
@@ -30,7 +30,7 @@ int main(int argc, char** argv) {
   cmd_opts opts = {
     .force = false
   };
-  if (!opt_parse(options, 5, &opts, argc - 1, argv + 1)) {
+  if (!opt_parse(options, 5, &opts, (size_t) argc - 1, argv + 1)) {
     fprintf(stderr, "error\n");
     return B_EXIT_CODE_USAGE_ERROR;
   }
